resolve conflicts

+++ src/entities/comment.js
@@ -1,6 +1,14 @@
 const Comment = require("../models/comment");
+const Post = require("../models/post");
+const User = require("./user");
+const Error404 = require("../middleware/error/error404");
+const Error400 = require("../middleware/error/error400");
 
 async function getCommentsByPost(postId) {
+  const post = await Post.query().findById(postId);
+  if (!post) {
+    throw new Error404("Post not exists");
+  }
   return Comment.query()
     .where({
       postId,
@@ -9,19 +17,35 @@ async function getCommentsByPost(postId) {
 }
 
 async function getCommentById(id) {
-  return Comment.query().findById(id);
+  const comment = await Comment.query().findById(id);
+  if (!comment) {
+    throw new Error404("Comment not exists");
+  }
+  return comment;
 }
 
 async function createComment(data) {
-  return Comment.query().insert(data);
+  const result = await User.getUserById(data.author);
+  if (!result) {
+    throw new Error400("User not exists");
+  }
+  await Comment.query().insert(data);
 }
 
 async function updateComment(id, data) {
-  return Comment.query().findById(id).patch(data);
+  const comment = await Comment.query().findById(id);
+  if (!comment) {
+    throw new Error404("Comment not exists");
+  }
+  await Comment.query().findById(id).patch(data);
 }
 
 async function removeComment(id) {
-  return Comment.query().findById(id).delete();
+  const comment = await Comment.query().findById(id);
+  if (!comment) {
+    throw new Error404("Comment not exists");
+  }
+  await Comment.query().findById(id).delete();
 }
 
 module.exports = {
